+++ src/lib/components/admin/user_utilities.js
@@ -1,8 +1,9 @@
 /**
  * @param {Array<import('$lib/types').User & {id: number}>} users
  * @param {number} currentAdminId
+ * @param {boolean|true=} prioritizeSuperusers
  */
-export function sortUsers(users, currentAdminId) {
+export function sortUsers(users, currentAdminId, prioritizeSuperusers = true) {
 	users.sort((a, b) => {
 		// current admin is always first
 		if (a.id === currentAdminId) {
@@ -11,12 +12,13 @@ export function sortUsers(users, currentAdminId) {
 		if (b.id === currentAdminId) {
 			return 1;
 		}
-		// prioritize superusers
-		if (a.is_superuser && !b.is_superuser) {
-			return -1;
-		}
-		if (!a.is_superuser && b.is_superuser) {
-			return 1;
+		if (prioritizeSuperusers) {
+			if (a.is_superuser && !b.is_superuser) {
+				return -1;
+			}
+			if (!a.is_superuser && b.is_superuser) {
+				return 1;
+			}
 		}
 		// then sort by email
 		return a.email < b.email ? -1 : 1;