+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Improved sorting of users in dropdown of the admin jobs page (\#402).
+* Fixed bug in retrieval of job log from the admin jobs page (\#402).
+* Highlighted relevant part of the error message in workflow job log modal (\#402).
+* Made the error message directly accessible from the new workflow page (\#402).
+
 # 0.9.0
 
 This release requires fractal-server 1.4.3.